+++ lib/processor/dedupe_test.go
@@ -323,6 +323,9 @@ func (e errCache) Get(key string) ([]byte, error) {
 func (e errCache) Set(key string, value []byte) error {
 	return errors.New("test err")
 }
+func (e errCache) SetMulti(items map[string][]byte) error {
+	return errors.New("test err")
+}
 func (e errCache) Add(key string, value []byte) error {
 	return errors.New("test err")
 }