+++ lib/types/interface.go
@@ -38,6 +38,10 @@ type Cache interface {
 	// fails.
 	Set(key string, value []byte) error
 
+	// SetMulti attempts to set the value of multiple keys, returns an error if
+	// any of the keys fail.
+	SetMulti(items map[string][]byte) error
+
 	// Add attempts to set the value of a key only if the key does not already
 	// exist, returns an error if the key already exists or if the command
 	// fails.